fix: 64bit stuff: int/pointer

+++ Singular/iparith.cc
@@ -5038,7 +5038,7 @@ static BOOLEAN jjTYPEOF(leftv res, leftv v)
 }
 static BOOLEAN jjUNIVARIATE(leftv res, leftv v)
 {
-  res->data=(char *)pIsUnivariate((poly)v->Data());
+  res->data=(char *)(long)pIsUnivariate((poly)v->Data());
   return FALSE;
 }
 static BOOLEAN jjVAR1(leftv res, leftv v)
@@ -6954,13 +6954,13 @@ static BOOLEAN jjLU_INVERSE(leftv res, leftv v)
   if (invertible)
   {
     ll->Init(2);
-    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)invertible;
+    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)(long)invertible;
     ll->m[1].rtyp=MATRIX_CMD; ll->m[1].data=(void *)iMat;
   }
   else
   {
     ll->Init(1);
-    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)invertible;
+    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)(long)invertible;
   }
 
   res->data=(char*)ll;
@@ -7038,14 +7038,14 @@ static BOOLEAN jjLU_SOLVE(leftv res, leftv v)
   if (solvable)
   {
     ll->Init(3);
-    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)solvable;
+    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)(long)solvable;
     ll->m[1].rtyp=MATRIX_CMD; ll->m[1].data=(void *)xVec;
     ll->m[2].rtyp=MATRIX_CMD; ll->m[2].data=(void *)homogSolSpace;
   }
   else
   {
     ll->Init(1);
-    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)solvable;
+    ll->m[0].rtyp=INT_CMD;    ll->m[0].data=(void *)(long)solvable;
   }
 
   res->data=(char*)ll;
