+++ Singular/misc_ip.cc
@@ -45,7 +45,7 @@ void setListEntry(lists L, int index, mpz_t n)
     if ((((ui<<3)>>3)==ui)
     && (mpz_cmp_si(n,(long)ui)==0))
     {
-      L->m[index].rtyp = INT_CMD; L->m[index].data = (void*)ui;
+      L->m[index].rtyp = INT_CMD; L->m[index].data = (void*)(long)ui;
       return;
     }
   }
@@ -59,7 +59,7 @@ void setListEntry_ui(lists L, int index, unsigned long ui)
   int i=(int)ui;
   if ((((unsigned long)i)==ui) && (((i<<3)>>3)==i))
   {
-    L->m[index].rtyp = INT_CMD; L->m[index].data = (void*)i;
+    L->m[index].rtyp = INT_CMD; L->m[index].data = (void*)(long)i;
   }
   else
   {
@@ -356,7 +356,7 @@ lists primeFactorisation(const number n, const int pBound)
   for (i = 0; i < index; i++)
   {
     multiplicitiesL->m[i].rtyp = INT_CMD;
-    multiplicitiesL->m[i].data = (void*)multiplicities[i];
+    multiplicitiesL->m[i].data = (void*)(long)multiplicities[i];
   }
   omFree(multiplicities);
